+++ API/controllers/concert.controller.js
@@ -3,6 +3,12 @@ const Concert = db.concert;
 const Op = db.Sequelize.Op;
 
 export const create = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour créer un concert"
+        });
+        return;
+    }
     if (!req.body.annee) {
         res.status(400).send({
             message: "Content can not be empty!"
@@ -31,8 +37,6 @@ export const create = (req, res) => {
 }
 
 export const findAll = (req, res) => {
-    const nom = req.query.nom;
-
     Concert.findAll(
         { include: [{model:db.artiste},
             {model:db.scene},
@@ -63,6 +67,12 @@ export const findOne = (req, res) => {
 }
 
 export const update = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour modifier un concert"
+        });
+        return;
+    }
     const id = req.params.id;
 
     Concert.update(req.body, {
@@ -87,6 +97,12 @@ export const update = (req, res) => {
 }
 
 export const deleteOne = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer un concert"
+        });
+        return;
+    }
     const id = req.params.id;
 
     Concert.destroy({
@@ -111,6 +127,12 @@ export const deleteOne = (req, res) => {
 }
 
 export const deleteAll = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer les concerts"
+        });
+        return;
+    }
     Concert.destroy({
         where: {},
         truncate: false
