+++ API/controllers/nationalite.controller.js
@@ -3,6 +3,13 @@ const Nationalite = db.nationalite;
 const Op = db.Sequelize.Op;
 
 export const create = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour créer une nationalité d'un artiste"
+        });
+        return;
+    }
+    
     if (!req.body.id_artiste) {
         res.status(400).send({
             message: "Content can not be empty!"
@@ -57,6 +64,13 @@ export const findOne = (req, res) => {
 }
 
 export const update = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour modifier une nationalité d'un artiste"
+        });
+        return;
+    }
+    
     const id = req.params.id;
 
     Nationalite.update(req.body, {
@@ -81,6 +95,13 @@ export const update = (req, res) => {
 }
 
 export const deleteByIdArtiste = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer une nationalité d'un artiste"
+        });
+        return;
+    }
+    
     const id = req.params.id;
 
     Nationalite.destroy({
@@ -97,6 +118,13 @@ export const deleteByIdArtiste = (req, res) => {
 }
 
 export const deleteAll = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer toutes les nationalitées d'un artiste"
+        });
+        return;
+    }
+    
     Nationalite.destroy({
         where: {},
         truncate: false
