+++ API/controllers/possedePrevious.controller.js
@@ -3,6 +3,13 @@ const Possede = db.possedePrevious;
 const Op = db.Sequelize.Op;
 
 export const create = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour créer un ancien lien entre un réseau social et un artiste"
+        });
+        return;
+    }
+    
     if (!req.body.lien) {
         res.status(400).send({
             message: "Content can not be empty!"
@@ -56,6 +63,13 @@ export const findOne = (req, res) => {
 }
 
 export const update = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour modifier un ancien lien entre un réseau social et un artiste"
+        });
+        return;
+    }
+    
     const id = req.params.id;
 
     Possede.update(req.body, {
@@ -80,6 +94,13 @@ export const update = (req, res) => {
 }
 
  export const deleteByIdArtiste = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer un ancien lien entre un réseau social et un artiste"
+        });
+        return;
+    }
+    
     const id = req.params.id;
 
     Possede.destroy({
@@ -98,6 +119,13 @@ export const update = (req, res) => {
 
 
  export const deleteAll =(req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer tous les anciens liens entre un réseau social et un artiste"
+        });
+        return;
+    }
+    
     Possede.destroy({
         where: {},
         truncate: false
