+++ API/controllers/reseauxSociaux.controller.js
@@ -3,6 +3,13 @@ const ReseauxSociaux = db.reseauxSociaux;
 const Op = db.Sequelize.Op;
 
 export const create = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour créer un réseau social"
+        });
+        return;
+    }
+    
     // Validate request
     if (!req.body.libelle) {
         res.status(400).send({
@@ -58,6 +65,13 @@ export const findOne = (req, res) => {
 }
 
 export const update = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour modifier un réseau social"
+        });
+        return;
+    }
+    
     const id = req.params.id;
     
     ReseauxSociaux.update(req.body, {
@@ -82,6 +96,13 @@ export const update = (req, res) => {
 }
 
 export const deleteOne = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer un réseau social"
+        });
+        return;
+    }
+    
     const id = req.params.id;
     
     ReseauxSociaux.destroy({
@@ -106,6 +127,13 @@ export const deleteOne = (req, res) => {
 }
 
 export const deleteAll = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer tous les réseaux sociaux"
+        });
+        return;
+    }
+    
     ReseauxSociaux.destroy({
         where: {},
         truncate: false
