+++ API/controllers/scene.controller.js
@@ -3,6 +3,13 @@ const Scene = db.scene;
 const Op = db.Sequelize.Op;
 
 export const create = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour créer une scène"
+        });
+        return;
+    }
+    
     if (!req.body.libelle) {
         res.status(400).send({
             message: "Content can not be empty!"
@@ -30,8 +37,6 @@ export const create = (req, res) => {
 }
 
 export const findAll = (req, res) => {
-    const nom = req.query.nom;
-
     Scene.findAll({ include: [{model: db.typescene}] })
         .then(data => {
             res.send(data);
@@ -45,6 +50,14 @@ export const findAll = (req, res) => {
 }
 
 export const findOne = (req, res) => {
+    
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour chercher une scène"
+        });
+        return;
+    }
+
     const id = req.params.id;
 
     Scene.findByPk(id, { include: [{model: db.typescene}] })
@@ -83,6 +96,13 @@ export const update = (req, res) => {
 }
 
 export const deleteOne = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer une scène"
+        });
+        return;
+    }
+    
     const id = req.params.id;
 
     Scene.destroy({
@@ -107,6 +127,13 @@ export const deleteOne = (req, res) => {
 }
 
 export const deleteAll = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer toutes les scènes"
+        });
+        return;
+    }
+    
     Scene.destroy({
         where: {},
         truncate: false
